requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
gspread==5.7.1
oauth2client==4.1.3
Werkzeug==2.3.7
pymongo==4.15.5
